PrismFlow/requirements.txt
# PrismFlow环境依赖文件 - 云端服务器验证版本
# 基于成功运行环境的精确版本配置

# 深度学习框架
torch==2.7.1
torchvision==0.22.1
triton==3.3.1

# Stable Diffusion相关
diffusers==0.29.2
transformers==4.48.3
accelerate==1.8.1
safetensors==0.5.3
tokenizers==0.21.2
huggingface-hub==0.24.7

# ControlNet和图像处理
controlnet_aux==0.0.10
timm==1.0.16
einops==0.8.1
peft==0.13.2

# 图像和视频处理
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
pillow==10.4.0
imageio==2.37.0
scikit-image==0.24.0

# 数值计算
numpy==1.26.4
scipy==1.13.1
matplotlib==3.9.4

# Web界面框架
gradio==3.50.2
gradio_client==0.6.1
fastapi==0.116.0
uvicorn==0.35.0

# 数据处理
pandas==2.3.1
pytz==2025.2

# 系统工具
tqdm==4.67.1
psutil==7.0.0
pydub==0.25.1
ffmpy==0.6.0

# 网络和HTTP
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
aiofiles==23.2.1

# 配置和验证
pydantic==1.10.22
PyYAML==6.0.2
click==8.1.8

# 可视化和UI
altair==5.5.0
rich==14.0.0

# 文件处理
filelock==3.18.0
fsspec==2025.5.1

# 数学和符号计算
sympy==1.14.0
mpmath==1.3.0

# 其他工具
regex==2024.11.6
typing_extensions==4.14.1
packaging==25.0
markdown-it-py==3.0.0

# NVIDIA CUDA支持包（如果使用GPU）
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77

# 开发工具
ruff==0.12.2 